+++ data/districts.json
@@ -1,29 +1,29 @@
 {
-  "0":	{ "name_en": "All electoral districts", "name_ar": "كافة الدوائر الانتخابية" },
-  "1":	{ "name_en": "Beirut 1", "name_ar": "بيروت الأولى" },
-  "2":	{ "name_en": "Beirut 2", "name_ar": "بيروت الثانية" },
-  "3":	{ "name_en": "Chouf", "name_ar": "الشوف" },
-  "4":	{ "name_en": "Metn", "name_ar": "المتن" },
-  "5":	{ "name_en": "Baabda", "name_ar": "بعبدا" },
-  "6":	{ "name_en": "Jbeil", "name_ar": "جبيل" },
-  "7":	{ "name_en": "Aley", "name_ar": "عاليه" },
-  "8":	{ "name_en": "Kesrouan", "name_ar": "كسروان" },
-  "9":	{ "name_en": "Batroun", "name_ar": "البترون" },
-  "10":	{ "name_en": "Danniyeh", "name_ar": "الضنية" },
-  "11":	{ "name_en": "Koura", "name_ar": "الكورة" },
-  "12":	{ "name_en": "Nabatieh", "name_ar": "النبطية" },
-  "13":	{ "name_en": "Bcharre", "name_ar": "بشري" },
-  "14":	{ "name_en": "Bint Jbeil", "name_ar": "بنت جبيل" },
-  "15":	{ "name_en": "Jezzine", "name_ar": "جزين" },
-  "16":	{ "name_en": "Zahle", "name_ar": "زحلة" },
-  "17":	{ "name_en": "Zgharta", "name_ar": "زغرتا" },
-  "18":	{ "name_en": "Sour", "name_ar": "صور" },
-  "19":	{ "name_en": "Akkar", "name_ar": "عكار" },
-  "20":	{ "name_en": "Saida Villages", "name_ar": "قرى صيدا" },
-  "21":	{ "name_en": "West Bekaa & Rachayya", "name_ar": "البقاع الغربي وراشيا" },
-  "22":	{ "name_en": "Baalbeck - Hermel", "name_ar": "بعلبك والهرمل" },
-  "23":	{ "name_en": "Marjayoun - Hasbaya", "name_ar": "مرجعيون وحاصبيا" },
-  "24":	{ "name_en": "Saida", "name_ar": "صيدا" },
-  "25":	{ "name_en": "Tripoli", "name_ar": "طرابلس" },
-  "26":	{ "name_en": "Minyeh", "name_ar": "المنية" }
+  "0":	{ "id": "all_districts", "name_en": "All electoral districts", "name_ar": "كافة الدوائر الانتخابية" },
+  "1":	{ "id": "beirut-1", "name_en": "Beirut 1", "name_ar": "بيروت الأولى" },
+  "2":	{ "id": "beirut-1", "name_en": "Beirut 2", "name_ar": "بيروت الثانية" },
+  "3":	{ "id": "chouf", "name_en": "Chouf", "name_ar": "الشوف" },
+  "4":	{ "id": "metn", "name_en": "Metn", "name_ar": "المتن" },
+  "5":	{ "id": "baabda", "name_en": "Baabda", "name_ar": "بعبدا" },
+  "6":	{ "id": "jbeil", "name_en": "Jbeil", "name_ar": "جبيل" },
+  "7":	{ "id": "aley", "name_en": "Aley", "name_ar": "عاليه" },
+  "8":	{ "id": "kesrouan", "name_en": "Kesrouan", "name_ar": "كسروان" },
+  "9":	{ "id": "batroun", "name_en": "Batroun", "name_ar": "البترون" },
+  "10":	{ "id": "danniyeh", "name_en": "Danniyeh", "name_ar": "الضنية" },
+  "11":	{ "id": "koura", "name_en": "Koura", "name_ar": "الكورة" },
+  "12":	{ "id": "nabatieh", "name_en": "Nabatieh", "name_ar": "النبطية" },
+  "13":	{ "id": "bcharre", "name_en": "Bcharre", "name_ar": "بشري" },
+  "14":	{ "id": "bintjbeil", "name_en": "Bint Jbeil", "name_ar": "بنت جبيل" },
+  "15":	{ "id": "jezzine", "name_en": "Jezzine", "name_ar": "جزين" },
+  "16":	{ "id": "zahle", "name_en": "Zahle", "name_ar": "زحلة" },
+  "17":	{ "id": "zgharta", "name_en": "Zgharta", "name_ar": "زغرتا" },
+  "18":	{ "id": "sour", "name_en": "Sour", "name_ar": "صور" },
+  "19":	{ "id": "akkar", "name_en": "Akkar", "name_ar": "عكار" },
+  "20":	{ "id": "saidavillages", "name_en": "Saida Villages", "name_ar": "قرى صيدا" },
+  "21":	{ "id": "westbekaa", "name_en": "West Bekaa & Rachayya", "name_ar": "البقاع الغربي وراشيا" },
+  "22":	{ "id": "baalbeck", "name_en": "Baalbeck - Hermel", "name_ar": "بعلبك والهرمل" },
+  "23":	{ "id": "marjayoun", "name_en": "Marjayoun - Hasbaya", "name_ar": "مرجعيون وحاصبيا" },
+  "24":	{ "id": "saida", "name_en": "Saida", "name_ar": "صيدا" },
+  "25":	{ "id":"tripoli", "name_en": "Tripoli", "name_ar": "طرابلس" },
+  "26":	{ "id": "minyeh", "name_en": "Minyeh", "name_ar": "المنية" }
 }
